Nodejs 12 EOL
The version 12.22.12 is planned to be the final Node.js 12 release. Node.js 12 will
reach End-of-Life status on 30 April 2022, after which it will no
receive updates. You are strongly advised to migrate your applications
to Node.js 16 or 14 (both of which are Long Term Support (LTS) releases)
to continue to receive future security updates beyond 30 April 2022.

https://nodejs.org/es/blog/release/v12.22.12

This commit update GitHub Action Workflow (CI) and information in README.md.